requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
aws-cdk.aws-ec2-alpha>=2.0.0a0
aws-cdk.aws-ssm-alpha>=2.0.0a0
